SiEPIC spice loading simplifications
