+++ examples/siepic/siepic.py
@@ -12,19 +12,13 @@ import numpy as np
 
 from simphony.netlist import Subcircuit
 from simphony.simulation import SweepSimulation, MonteCarloSweepSimulation
-from simphony.plugins.siepic.builders import build_circuit
-from simphony.plugins.siepic.parser import load_spi
-
+from simphony.plugins.siepic import load
 
 # The siepic plugin only needs the path to the main spice file.
 filename = os.path.join(os.path.dirname(__file__), 'MZI4', 'MZI4_main.spi')
 
 # Loading the main file includes any other files linked to it internally.
-data = load_spi(filename)
-
-# Building the circuit returns the subcircuits, the circuits, and the 
-# simulation objects.
-built = build_circuit(data, 'simphony.library.siepic')
+built = load(filename)
 
 # The file `MZI4_main.spi` declares one single-sweep simulation.
 analyses = built['analyses']
@@ -38,7 +32,6 @@ sim.circuit.pins
 # take the data from the ports, using the pin names we looked up earlier.
 res = sim.simulate()
 f, s = res.data('ebeam_gc_te1550_laser1', 'ebeam_gc_te1550_detector2')
-# plt.plot(f*1e9, s)
 plt.plot(f, s)
 plt.title("MZI")
 plt.tight_layout()
